refactor(core): 拆分上下文压缩编排边界
