+++ TODO.md
@@ -9,7 +9,7 @@
 - [x] 自动压缩和手动压缩支持 Cancel/Shutdown。
 - [x] 当前任务续接只进入下一次模型请求，不写入 Session 或前端消息。
 - [x] 压缩成功、失败和取消均发布完整界面反馈。
-- [x] 压缩任务启动、结果提交、用量和通知统一由 ContextCompressor 处理，执行循环只推进状态。
+- [x] ContextCompressor 只负责生成摘要，react/compression 统一处理任务生命周期、结果提交、用量和通知。
 - [x] 压缩用量进入本轮最终用量，压缩后的当前 token 使用实际输出校准。
 - [x] 补齐预算、截断、一次性续接、取消和失败原子性测试。
 - [x] 通过 Rust 检查、相关测试和前端构建。
