+++ crates/tiangong-core/src/context/compressor.rs
@@ -1,10 +1,5 @@
-use anyhow::{Result, bail};
-
-use crate::context::organizer::ContextOrganizer;
 use crate::model::{ModelRequest, SingleProviderClient, StopReason, TokenUsage};
-use crate::session::{ContentBlock, Message, MessagePhase, MessageRole, Session};
-use crate::turn_context::TurnContext;
-use tiangong_types::{StreamEvent, stream::ContextCompressAction};
+use crate::session::{Message, MessagePhase, MessageRole, Session};
 
 /// 判断消息是否为可压缩的有效消息。
 fn is_compressible(message: &Message) -> bool {
@@ -13,7 +8,7 @@ fn is_compressible(message: &Message) -> bool {
         && message.phase != MessagePhase::CompressedResume
 }
 
-/// 使用 Session 和客户端快照执行压缩请求，并统一处理任务启动、结果提交和通知。
+/// 使用 Session 和客户端快照生成上下文摘要，不负责任务调度、持久化或通知。
 pub struct ContextCompressor {
     session: Session,
     client: SingleProviderClient,
@@ -60,11 +55,8 @@ impl std::fmt::Display for CompressionError {
 impl std::error::Error for CompressionError {}
 
 impl ContextCompressor {
-    pub fn new(ctx: &TurnContext) -> Self {
-        Self {
-            session: ctx.session.clone(),
-            client: ctx.client.clone(),
-        }
+    pub fn new(session: Session, client: SingleProviderClient) -> Self {
+        Self { session, client }
     }
 
     pub fn has_pending_messages(&self) -> bool {
@@ -125,229 +117,6 @@ impl ContextCompressor {
             summary_up_to: split_point,
             boundary_message_id,
         })
-    }
-
-    pub(crate) fn start_auto(
-        ctx: &TurnContext,
-        organizer: &ContextOrganizer,
-        observed_tokens: usize,
-        force_target_budget: bool,
-    ) -> tokio::task::JoinHandle<std::result::Result<CompressionUpdate, CompressionError>> {
-        Self::notify_started(ctx);
-        let budget_tokens = if force_target_budget {
-            0
-        } else {
-            observed_tokens
-        };
-        Self::start_task(Self::new(ctx), organizer, budget_tokens, true)
-    }
-
-    pub(crate) fn start_manual(
-        ctx: &TurnContext,
-        organizer: &ContextOrganizer,
-        observed_tokens: usize,
-    ) -> Option<tokio::task::JoinHandle<std::result::Result<CompressionUpdate, CompressionError>>>
-    {
-        Self::notify_started(ctx);
-        let compressor = Self::new(ctx);
-        if !compressor.has_pending_messages() {
-            Self::notify_result(ctx, ContextCompressAction::Noop);
-            return None;
-        }
-        Some(Self::start_task(
-            compressor,
-            organizer,
-            observed_tokens,
-            false,
-        ))
-    }
-
-    fn start_task(
-        compressor: Self,
-        organizer: &ContextOrganizer,
-        observed_tokens: usize,
-        include_current_task: bool,
-    ) -> tokio::task::JoinHandle<std::result::Result<CompressionUpdate, CompressionError>> {
-        let output_budget = organizer.compression_output_budget(observed_tokens);
-        tokio::spawn(async move {
-            let output_budget = output_budget.ok_or_else(|| {
-                CompressionError::new("上下文剩余空间不足 2048 tokens，无法生成有效摘要")
-            })?;
-            compressor
-                .compress(output_budget, include_current_task)
-                .await
-        })
-    }
-
-    fn notify_started(ctx: &TurnContext) {
-        let _ = ctx.stream_tx.send(StreamEvent::ContextCompressing {
-            summary_up_to: ctx.session.summary_up_to,
-            total_messages: ctx.session.messages.len(),
-        });
-    }
-
-    pub(crate) fn notify_cancelled(ctx: &TurnContext) {
-        Self::notify_result(ctx, ContextCompressAction::Cancelled);
-    }
-
-    pub(crate) fn notify_cleared(
-        stream_tx: &std::sync::mpsc::Sender<StreamEvent>,
-        session: &Session,
-    ) {
-        Self::notify_session_result(stream_tx, session, ContextCompressAction::Clear);
-    }
-
-    fn notify_auto_success(
-        ctx: &TurnContext,
-        update: &CompressionUpdate,
-        current_tokens: usize,
-        observed_tokens: usize,
-    ) {
-        Self::notify_result(ctx, ContextCompressAction::Auto);
-        Self::notify_usage(ctx, &update.usage, Some(current_tokens), "context_summary");
-        tracing::info!(
-            session_id = %ctx.session.id,
-            observed_tokens,
-            old_summary_up_to = update.previous_summary_up_to,
-            summary_up_to = ctx.session.summary_up_to,
-            total_messages = update.summary_up_to,
-            "上下文摘要已更新"
-        );
-    }
-
-    fn notify_auto_failure(ctx: &TurnContext, usage: &TokenUsage, error: &dyn std::fmt::Display) {
-        Self::notify_usage(ctx, usage, None, "context_summary_failed");
-        tracing::warn!(
-            session_id = %ctx.session.id,
-            error = %error,
-            "上下文压缩失败，保留原始上下文"
-        );
-        Self::notify_result(ctx, ContextCompressAction::Failed);
-    }
-
-    fn notify_manual_success(ctx: &TurnContext, update: &CompressionUpdate, current_tokens: usize) {
-        Self::notify_result(ctx, ContextCompressAction::Compress);
-        Self::notify_usage(
-            ctx,
-            &update.usage,
-            Some(current_tokens),
-            "manual_context_compress",
-        );
-        tracing::info!(
-            session_id = %ctx.session.id,
-            summary_up_to = ctx.session.summary_up_to,
-            "手动上下文摘要已更新"
-        );
-    }
-
-    fn notify_manual_failure(ctx: &TurnContext, usage: &TokenUsage, error: &dyn std::fmt::Display) {
-        Self::notify_usage(ctx, usage, None, "manual_context_compress_failed");
-        tracing::warn!(
-            session_id = %ctx.session.id,
-            error = %error,
-            "手动上下文压缩失败，继续使用原始上下文"
-        );
-        let _ = ctx.stream_tx.send(StreamEvent::AgentNotification {
-            agent_id: "system".to_string(),
-            agent_label: "系统".to_string(),
-            content: format!("手动压缩上下文失败：{error}"),
-            level: "error".to_string(),
-        });
-        Self::notify_result(ctx, ContextCompressAction::Failed);
-    }
-
-    fn notify_usage(
-        ctx: &TurnContext,
-        usage: &TokenUsage,
-        current_tokens: Option<usize>,
-        source: &'static str,
-    ) {
-        crate::react::context::emit_token_usage(
-            &ctx.stream_tx,
-            usage,
-            current_tokens,
-            ctx.context_limit,
-            source,
-            None,
-        );
-    }
-
-    pub(crate) fn complete_auto(
-        ctx: &mut TurnContext,
-        accumulated_usage: &mut TokenUsage,
-        observed_tokens: usize,
-        result: std::result::Result<
-            std::result::Result<CompressionUpdate, CompressionError>,
-            tokio::task::JoinError,
-        >,
-    ) -> Option<Message> {
-        let result = result.unwrap_or_else(|error| Err(CompressionError::new(error.to_string())));
-
-        match result {
-            Ok(update) => {
-                accumulated_usage.accumulate(&update.usage);
-                match apply_compression(ctx, &update, false) {
-                    Ok(current_tokens) => {
-                        Self::notify_auto_success(ctx, &update, current_tokens, observed_tokens);
-                        update
-                            .current_task
-                            .as_deref()
-                            .map(build_compression_resume_message)
-                    }
-                    Err(error) => {
-                        Self::notify_auto_failure(ctx, &update.usage, &error);
-                        None
-                    }
-                }
-            }
-            Err(error) => {
-                accumulated_usage.accumulate(&error.usage);
-                Self::notify_auto_failure(ctx, &error.usage, &error);
-                None
-            }
-        }
-    }
-
-    pub(crate) fn complete_manual(
-        ctx: &mut TurnContext,
-        result: std::result::Result<
-            std::result::Result<CompressionUpdate, CompressionError>,
-            tokio::task::JoinError,
-        >,
-    ) {
-        let result = result.unwrap_or_else(|error| Err(CompressionError::new(error.to_string())));
-
-        match result {
-            Ok(update) => match apply_compression(ctx, &update, true) {
-                Ok(current_tokens) => Self::notify_manual_success(ctx, &update, current_tokens),
-                Err(error) => {
-                    ctx.session.token_usage.accumulate(&update.usage);
-                    ctx.session.persist_to_disk();
-                    Self::notify_manual_failure(ctx, &update.usage, &error);
-                }
-            },
-            Err(error) => {
-                ctx.session.token_usage.accumulate(&error.usage);
-                ctx.session.persist_to_disk();
-                Self::notify_manual_failure(ctx, &error.usage, &error);
-            }
-        }
-    }
-
-    fn notify_result(ctx: &TurnContext, action: ContextCompressAction) {
-        Self::notify_session_result(&ctx.stream_tx, &ctx.session, action);
-    }
-
-    fn notify_session_result(
-        stream_tx: &std::sync::mpsc::Sender<StreamEvent>,
-        session: &Session,
-        action: ContextCompressAction,
-    ) {
-        let _ = stream_tx.send(StreamEvent::ContextCompressed {
-            action,
-            summary_up_to: session.summary_up_to,
-            remaining_messages: session.messages.len().saturating_sub(session.summary_up_to),
-        });
     }
 
     fn summary_request(
@@ -470,59 +239,6 @@ impl ContextCompressor {
     }
 }
 
-/// 将压缩结果应用到候选 Session，持久化成功后再替换真实 Session。
-pub fn apply_compression(
-    ctx: &mut TurnContext,
-    update: &CompressionUpdate,
-    account_usage_in_session: bool,
-) -> Result<usize> {
-    if ctx.session.summary_up_to != update.previous_summary_up_to {
-        bail!(
-            "压缩期间摘要边界已变化：expected={}, actual={}",
-            update.previous_summary_up_to,
-            ctx.session.summary_up_to
-        );
-    }
-    let Some(boundary) = update
-        .summary_up_to
-        .checked_sub(1)
-        .and_then(|index| ctx.session.messages.get(index))
-    else {
-        bail!("压缩结果边界无效：{}", update.summary_up_to);
-    };
-    if boundary.id != update.boundary_message_id {
-        bail!("压缩期间消息边界已变化，拒绝提交过期结果");
-    }
-
-    let mut candidate = ctx.session.clone();
-    candidate.context_summary = Some(update.summary.clone());
-    candidate.summary_up_to = update.summary_up_to;
-    mark_compact_boundary(&mut candidate.messages, update.summary_up_to);
-    if account_usage_in_session {
-        candidate.token_usage.accumulate(&update.usage);
-        candidate.active_agent_current_tokens = 0;
-        candidate.agent_current_tokens.clear();
-    }
-    let current_tokens = update.usage.completion_tokens;
-    candidate.current_tokens = current_tokens;
-    crate::react::context::rebuild_system_prompt_for_session(&mut candidate, &ctx.plugins);
-    candidate
-        .try_persist_to_disk()
-        .map_err(anyhow::Error::msg)?;
-    ctx.session = candidate;
-    Ok(current_tokens)
-}
-
-/// 构造只进入下一次模型请求的任务续接消息。
-pub fn build_compression_resume_message(current_task: &str) -> Message {
-    let mut message = Message::new(MessageRole::User, "");
-    message.content = vec![ContentBlock::model_instruction(format!(
-        "以下是上下文压缩后的当前任务状态，请据此继续执行，不要重新询问用户：\n\n{current_task}"
-    ))];
-    message.phase = MessagePhase::Normal;
-    message
-}
-
 pub fn mark_compact_boundary(messages: &mut [Message], split_point: usize) {
     for message in messages.iter_mut() {
         message.compact = false;
@@ -605,18 +321,5 @@ mod tests {
 
         assert_eq!(summary, "普通摘要");
         assert!(task.is_none());
-    }
-
-    #[test]
-    fn resume_message_is_transient_normal_phase() {
-        let message = build_compression_resume_message("继续任务");
-
-        assert_eq!(message.role, MessageRole::User);
-        assert_eq!(message.phase, MessagePhase::Normal);
-        assert!(!message.model_excluded);
-        assert!(matches!(
-            &message.content[0],
-            ContentBlock::ModelInstruction { text } if text.contains("继续任务")
-        ));
     }
 }
